backend/requirements.txt
Django>=4.0,<5.0
requests
django-cors-headers
djangorestframework
djangorestframework-simplejwt